backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.7
Deprecated==1.3.1
fastapi==0.135.3
firebase_admin==7.4.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-logging==3.15.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
limits==5.8.0
msgpack==1.1.2
opentelemetry-api==1.41.0
packaging==26.1
pluggy==1.6.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-cov==7.1.0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
slowapi==0.1.9
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
zipp==3.23.1
